[refactor] move delegation editor and proof editor to own modules
Split the avalanche_dialogs classes in submodules of a `qt.avalanche` package.
